fix: Support `dataset.set_owner(None)`

+++ uvdat/core/models/dataset.py
@@ -62,12 +62,13 @@ class Dataset(models.Model):
             permission__codename="owner",
         ).delete()
 
-        # Delete any existing user perms and set owner
-        UserObjectPermission.objects.filter(
-            **filters,
-            user_id__in=[user.id],
-        ).delete()
-        assign_perm("owner", user, self)
+        if user is not None:
+            # Delete any existing user perms and set owner
+            UserObjectPermission.objects.filter(
+                **filters,
+                user_id__in=[user.id],
+            ).delete()
+            assign_perm("owner", user, self)
 
     @transaction.atomic()
     def set_tags(self, tags: list[str]):
